requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pillow==10.0.1
requests==2.31.0
sentry-sdk==1.32.0
google-generativeai==0.3.1
python-dotenv
